+++ myelin/kernel/sse-matmul.cc
@@ -204,8 +204,8 @@ class SSEFltVecMatMulBase : public Kernel {
     __ j(less, &l1);
   }
 
-  int Complexity(const Step *step) override {
-    int ops = step->input(1)->elements() * 2;
+  int64 Complexity(const Step *step) override {
+    int64 ops = step->input(1)->elements() * 2;
     if (bias_) ops += step->input(2)->elements();
     if (relu_) ops += step->output(0)->elements();
     return ops;
